Neural Networks
I have posted some basic configuration code for the first versions of the neural network that will be used, on Google Compute, to find prototypes and complete patterns. These Artificial Neural Networks are a form of Artificial Intelligence that will grow more sophisticated as the visitor interacts with it. This basic form will be scaled up and modified to embrace the characteristics of Quantum Neural Networks, that analyse numerical input according to a 'Holographic' method. Here the visitor is interacting with an AI Holographic system that will work to produce the visitor 'Art of the Multiverse' output, showing connections between our minds/cognition and the Quantum Universe.